==> Project5_Predicting_Hard_Drive_Failure/1_GetData.ipynb <==
import pandas as pd
import zipfile
import os
from os import listdir, walk
from os.path import isfile, join
import re

import matplotlib.pyplot as plt


# we access the name of each file, attach it to the path and save
# in a list
mypath = '/Users/aminenhila/Desktop/Metis/Project5/Data/2019/'
filenames_list = []
file_path_list = []
dirpath_list = []
for (dirpath, dirnames, filenames) in walk(mypath):
    if filenames !=[]:
        filenames = [f for f in filenames if not f[0] == '.']
        dirnames[:] = [d for d in dirnames if not d[0] == '.']
        filenames_list.extend(filenames)
        file_path = [dirpath+'/'+f for f in filenames]
        file_path_list.extend(file_path)

file_path_list = sorted(file_path_list)

def get_hd_data(file_path_list, columns):
    ''' This function takes in a list of file paths,
    the names of columns (raw features) we want,
    and returns all the daily info on the hard drives
    '''
    hard_drive_df = pd.DataFrame()
    # we don't need the normalized columns
    # don't need most of the variables
    for file in file_path_list:
        SGate_4_df = pd.read_csv(file)
        SGate_4_df = SGate_4_df[SGate_4_df.model == 'ST4000DM000'][columns]
        hard_drive_df = pd.concat([hard_drive_df, SGate_4_df],axis = 0, ignore_index=True)
    return hard_drive_df


# we only want raw features and only 12 of them
# the more critical ones for failure diagnostics
# according to Backblaze and wikipedia
columns = ['date', 'serial_number', 'model','capacity_bytes', 'failure',
            'smart_184_raw', 'smart_10_raw', 'smart_242_raw',
           'smart_241_raw','smart_194_raw', 'smart_193_raw', 'smart_9_raw',
           'smart_198_raw', 'smart_197_raw', 'smart_188_raw', 'smart_187_raw',
           'smart_5_raw']


# hard drives from 2019
Sgate_4_hard_drive_df = get_hd_data(file_path_list, columns)

Sgate_4_hard_drive_df

# save the file 
Sgate_4_hard_drive_df.to_csv('Sgate_4_hard_drive_df.csv')
