==> Project5_Predicting_Hard_Drive_Failure/2_CleanData.ipynb <==
This notebook cleans the data and prepares it for EDA


import pandas as pd

# read the data that we got from GetData notebook
Sgate_4_hard_drive_df = pd.read_csv('Sgate_4_hard_drive_df.csv', index_col=[0])

Sgate_4_hard_drive_df.head()

Sgate_4_hard_drive_df.info(null_counts=True)

There are 278 rows that contain no information on the SMART variables. 
We will replace the nan values with the prev_day_value+(rev_day_value-prev_prev_day_value) to avoid using future information in the current observation which can cause data leakage

Sgate_4_hard_drive_df.describe()

Sgate_4_hard_drive_df.capacity_bytes.value_counts()

Sgate_4_hard_drive_df[Sgate_4_hard_drive_df.capacity_bytes==-1]

We will not be using the capacity_bytes in our modeling since most of the hard drives have the same capacity, and all the failed ones except one have 4Tb capacity

We can see that smart_10 is all zeroes for all the hard drives so we will drop it

Sgate_4_hard_drive_df = Sgate_4_hard_drive_df.drop('smart_10_raw', axis = 1)

For smart_9, which is the number of hours that a hard drive has been operational, the max number is equivalent to about 6 years which might be a on the higher side of life expectancy of hard-drives, but we won't change anything for now

# convert the date column to datetime
Sgate_4_hard_drive_df.date = pd.to_datetime(Sgate_4_hard_drive_df.date)

# sort the df by serial number then by date
Sgate_4_hard_drive_df = Sgate_4_hard_drive_df.sort_values(['serial_number','date'], ascending=True).reset_index()

Sgate_4_hard_drive_df

Sgate_4_hard_drive_df.columns

Let's fix the nan values

# this gives us the serial numbers of the hard drive models that have 
# daily observation that has nans in it
serial_numbers_with_null= Sgate_4_hard_drive_df[Sgate_4_hard_drive_df.smart_197_raw.isnull()].serial_number.unique()
serial_numbers_with_null

mask = ['capacity_bytes','smart_184_raw', 'smart_242_raw', 'smart_241_raw', 'smart_194_raw',
       'smart_193_raw', 'smart_9_raw', 'smart_198_raw', 'smart_197_raw',
       'smart_188_raw', 'smart_187_raw', 'smart_5_raw']

for s_num in serial_numbers_with_null:
    df = Sgate_4_hard_drive_df[Sgate_4_hard_drive_df.serial_number == s_num]
    df
    nan_index = df[pd.isnull(df).any(axis = 1)].index
    # if the nan value is the first or second entry for the hard drive, then we will just
    # drop it
    # if it is 3rd or later then we will use the change between the previous 2 to add to
    # the prev value to replace the nan
    if len(nan_index) >1:
        print(s_num)
    for ind in nan_index:
        if ind in df.index[:2]:
            Sgate_4_hard_drive_df = Sgate_4_hard_drive_df.drop(index = ind)
        else:
            # take the different between the prev 2 values and add to the prev
            Sgate_4_hard_drive_df.loc[ind,mask] = df.loc[ind-1,mask] + (df.loc[ind-1,mask] - df.loc[ind-2,mask])
            df.loc[ind,mask] = df.loc[ind-1,mask] + (df.loc[ind-1,mask] - df.loc[ind-2,mask])


# check if worked
Sgate_4_hard_drive_df[Sgate_4_hard_drive_df.smart_197_raw.isnull()].serial_number

Sgate_4_hard_drive_df = Sgate_4_hard_drive_df.reset_index()

Sgate_4_hard_drive_df = Sgate_4_hard_drive_df.drop(['level_0', 'index'], axis = 1)

Sgate_4_hard_drive_df

Save the cleaned dataframe

Sgate_4_hard_drive_df.to_csv('cleaned_Sgate_4_hard_drive_df.csv', index=False)
