==> Project5_Predicting_Hard_Drive_Failure/4_CreateNewFeatures.ipynb <==
This notebook adds the following features to the hard drive dataframe:
1. moving average of smart values in the most recent w cycles
2. the standard deviation of sensor values in the most w recent cycles
3. velocity of change
4. The distance of observation from the mean of recent w observations   

import pandas as pd
import matplotlib.pyplot as plt
#import seaborn as sns

%matplotlib inline 


cleaned_Sgate_4_hard_drive_df = pd.read_csv('cleaned_Sgate_4_hard_drive_df.csv')

cleaned_Sgate_4_hard_drive_df.date = pd.to_datetime(cleaned_Sgate_4_hard_drive_df.date)

cleaned_Sgate_4_hard_drive_df

We won't use the capacity bites (since the hard drives have mainly the same capacity) and the smart_9_raw feature which is the number of operational hours 

cleaned_Sgate_4_hard_drive_df = cleaned_Sgate_4_hard_drive_df.drop(['capacity_bytes'],axis=1)

cleaned_Sgate_4_hard_drive_df.columns

mask = ['smart_184_raw', 'smart_242_raw', 'smart_241_raw',
       'smart_194_raw', 'smart_193_raw', 'smart_198_raw',
       'smart_197_raw', 'smart_188_raw', 'smart_187_raw', 'smart_5_raw']
mask_mean_rw_10 = ['smart_184_rw10_mean','smart_242_rw10_mean','smart_241_rw10_mean',
       'smart_194_rw10_mean', 'smart_193_rw10_mean', 'smart_198_rw10_mean',
       'smart_197_rw10_mean','smart_188_rw10_mean', 'smart_187_rw10_mean','smart_5_rw10_mean']
mask_mean_rw_20 = ['smart_184_rw20_mean','smart_242_rw20_mean','smart_241_rw20_mean',
       'smart_194_rw20_mean', 'smart_193_rw20_mean', 'smart_198_rw20_mean',
       'smart_197_rw20_mean','smart_188_rw20_mean', 'smart_187_rw20_mean','smart_5_rw20_mean']
mask_mean_rw_30 = ['smart_184_rw30_mean','smart_242_rw30_mean','smart_241_rw30_mean',
       'smart_194_rw30_mean', 'smart_193_rw30_mean', 'smart_198_rw30_mean',
       'smart_197_rw30_mean','smart_188_rw30_mean', 'smart_187_rw30_mean','smart_5_rw30_mean']
mask_std_rw_10 = ['smart_184_rw10_std','smart_242_rw10_std','smart_241_rw10_std',
       'smart_194_rw10_std', 'smart_193_rw10_std', 'smart_198_rw10_std',
       'smart_197_rw10_std','smart_188_rw10_std', 'smart_187_rw10_std','smart_5_rw10_std']
mask_std_rw_20 = ['smart_184_rw20_std','smart_242_rw20_std','smart_241_rw20_std',
       'smart_194_rw20_std', 'smart_193_rw20_std', 'smart_198_rw20_std',
       'smart_197_rw20_std','smart_188_rw20_std', 'smart_187_rw20_std','smart_5_rw20_std']
mask_std_rw_30 = ['smart_184_rw30_std','smart_242_rw30_std','smart_241_rw30_std',
       'smart_194_rw30_std', 'smart_193_rw30_std', 'smart_198_rw30_std',
       'smart_197_rw30_std','smart_188_rw30_std', 'smart_187_rw30_std','smart_5_rw30_std']
mask_vel = ['smart_184_vel', 'smart_242_vel', 'smart_241_vel',
       'smart_194_vel', 'smart_193_vel', 'smart_198_vel',
       'smart_197_vel', 'smart_188_vel', 'smart_187_vel', 'smart_5_vel']
mask_from_mean_10 = ['smart_184_from_mean_10', 'smart_242_from_mean_10', 'smart_241_from_mean_10',
       'smart_194_from_mean_10', 'smart_193_from_mean_10', 'smart_198_from_mean_10',
       'smart_197_from_mean_10', 'smart_188_from_mean_10', 'smart_187_from_mean_10', 'smart_5_from_mean_10']
mask_from_mean_20 = ['smart_184_from_mean_20', 'smart_242_from_mean_20', 'smart_241_from_mean_20',
       'smart_194_from_mean_20', 'smart_193_from_mean_20', 'smart_198_from_mean_20',
       'smart_197_from_mean_20', 'smart_188_from_mean_20', 'smart_187_from_mean_20', 'smart_5_from_mean_20']
mask_from_mean_30 = ['smart_184_from_mean_30', 'smart_242_from_mean_30', 'smart_241_from_mean_30',
       'smart_194_from_mean_30', 'smart_193_from_mean_30', 'smart_198_from_mean_30',
       'smart_197_from_mean_30', 'smart_188_from_mean_30', 'smart_187_from_mean_30', 'smart_5_from_mean_30']

#### 1. moving average of smart values in the most recent w cycles


def moving_average(mask,mask_mean, n):
    grouped_df = cleaned_Sgate_4_hard_drive_df.groupby(['serial_number'])[mask].rolling(window=n).mean().reset_index()
    grouped_df = grouped_df.rename(columns = {key:val for (key,val) in zip(mask, mask_mean)})
    grouped_df = grouped_df.drop(['level_1','serial_number'], axis = 1)
    return grouped_df

cleaned_Sgate_4_hard_drive_df = pd.concat([cleaned_Sgate_4_hard_drive_df, moving_average(mask,mask_mean_rw_10,10)],axis = 1)

cleaned_Sgate_4_hard_drive_df = pd.concat([cleaned_Sgate_4_hard_drive_df, moving_average(mask,mask_mean_rw_20,20)],axis = 1)

cleaned_Sgate_4_hard_drive_df = pd.concat([cleaned_Sgate_4_hard_drive_df, moving_average(mask,mask_mean_rw_30,30)],axis = 1)

cleaned_Sgate_4_hard_drive_df.columns

#### 2. the standard deviation of sensor values in the most w recent cycles


def moving_std(mask,mask_std,n):
    grouped_df = cleaned_Sgate_4_hard_drive_df.groupby(['serial_number'])[mask].rolling(window=n).std().reset_index()
    grouped_df = grouped_df.rename(columns = {key:val for (key,val) in zip(mask, mask_std)})
    grouped_df = grouped_df.drop(['level_1','serial_number'], axis = 1)
    return grouped_df

cleaned_Sgate_4_hard_drive_df = pd.concat([cleaned_Sgate_4_hard_drive_df, moving_std(mask,mask_std_rw_10,10)],axis = 1)

cleaned_Sgate_4_hard_drive_df = pd.concat([cleaned_Sgate_4_hard_drive_df, moving_std(mask,mask_std_rw_20,20)],axis = 1)

cleaned_Sgate_4_hard_drive_df = pd.concat([cleaned_Sgate_4_hard_drive_df, moving_std(mask,mask_std_rw_30,30)],axis = 1)

cleaned_Sgate_4_hard_drive_df.columns

#### 3. velocity of change

grouped_df = cleaned_Sgate_4_hard_drive_df.groupby(['serial_number'])[mask].diff().reset_index()
day_diff = cleaned_Sgate_4_hard_drive_df.groupby(['serial_number'])[['date']].diff().reset_index().date.dt.days
div = grouped_df.divide(day_diff, axis = 0)

div

div = div.rename(columns = {key:val for (key,val) in zip(mask, mask_vel)})


div = div.drop('index', axis = 1)
div

div.columns

cleaned_Sgate_4_hard_drive_df = pd.concat([cleaned_Sgate_4_hard_drive_df, div],axis = 1)

cleaned_Sgate_4_hard_drive_df.columns

df = cleaned_Sgate_4_hard_drive_df

#### 4. The distance of observation from the mean of recent w observations  

def dist_from_mean(mask,mask_mean,mask_from_mean):
    for raw,sm_mean, from_mean in zip(mask, mask_mean, mask_from_mean):
        cleaned_Sgate_4_hard_drive_df[from_mean] = cleaned_Sgate_4_hard_drive_df[raw]-cleaned_Sgate_4_hard_drive_df[sm_mean]
    

dist_from_mean(mask, mask_mean_rw_10, mask_from_mean_10)

dist_from_mean(mask, mask_mean_rw_20, mask_from_mean_20)

dist_from_mean(mask, mask_mean_rw_30, mask_from_mean_30)

print(cleaned_Sgate_4_hard_drive_df.columns[:50])
print(cleaned_Sgate_4_hard_drive_df.columns[50:])

# save the dataframe
cleaned_Sgate_4_hard_drive_df.to_csv('features_cleaned_Sgate_4_hard_drive_df.csv', index=False)
