==> Project5_Predicting_Hard_Drive_Failure/6_Regression.ipynb <==
This notebook applies regression to the hard drive data and tries to predict the number of days it takes for a hard drive to fail

import pandas as pd
import numpy as np
import random

hard_drive_df = pd.read_csv('failed_hard_drive_df.csv')

hard_drive_df.date = pd.to_datetime(hard_drive_df.date)

hard_drive_df.columns

hard_drive_df = hard_drive_df[['date','remaining_useful_life',
       'RUL_30', 'failure', 'model',
       'serial_number','smart_187_from_mean_10', 'smart_187_raw', 'smart_187_rw10_mean',
       'smart_187_rw10_std', 'smart_187_vel',
       'smart_188_from_mean_10', 'smart_188_raw', 'smart_188_rw10_mean',
       'smart_188_rw10_std', 'smart_188_vel','smart_197_from_mean_10', 'smart_197_raw', 'smart_197_rw10_mean',
       'smart_197_rw10_std', 'smart_197_vel',
       'smart_198_from_mean_10', 'smart_198_raw', 'smart_198_rw10_mean',
       'smart_198_rw10_std', 'smart_198_vel','smart_5_from_mean_10',
       'smart_5_raw', 'smart_5_rw10_mean', 'smart_5_rw10_std','smart_5_vel']]

hard_drive_df.info(null_counts=True)

The hard drive data contains some nan values created in the process of creating new features (some will have 10 nans per hard drive info, some 20, some 30, etc). Depending on the features used we will remove the coresponding rows containing nans

hard_drive_df = hard_drive_df.dropna(how='any').reset_index()   
hard_drive_df = hard_drive_df.drop('index', axis=1)

hard_drive_df.info()

len(hard_drive_df[hard_drive_df.failure == 1].serial_number.unique())

hard_drive_df.columns

from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor,GradientBoostingRegressor
from sklearn.linear_model import LinearRegression, Ridge, Lasso, ElasticNet,SGDRegressor
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import mean_squared_error, r2_score
import statsmodels.api as sm
from sklearn.neighbors import KNeighborsRegressor
from imblearn.over_sampling import RandomOverSampler

# 25 % test and 75 % train
failed_hard_drive_list = hard_drive_df[hard_drive_df.failure == 1].serial_number.unique()

# split the failed and working hard drive into 25% and 75% to use for train and test set
train_hd_failed, test_hd_failed = train_test_split(failed_hard_drive_list,test_size=0.25, random_state=42)


# get the train and test dataframes
train_df = hard_drive_df[hard_drive_df.serial_number.isin(train_hd_failed)]
test_df = hard_drive_df[hard_drive_df.serial_number.isin(test_hd_failed)]

train_df.sample(frac=1)
test_df.sample(frac=1)


train_df['remaining_useful_life'].std()

# split the train into 3 groups (3 fold)
kfold_5 = np.array_split(train_hd_failed,3)

mask = ['RUL_30','smart_5_raw','smart_187_raw','smart_188_raw','smart_197_raw','smart_198_raw']


train_df.columns

r2_score_list_tr = []
r2_score_list_vl = []
for val_group_index in range(len(kfold_5)):
    # get serial numbers in val and train sets
    train_s_nums = kfold_5[:val_group_index]+kfold_5[val_group_index+1:]
    val_s_nums = kfold_5[val_group_index]

    # get the data for train and val sets
    train_data = train_df[train_df.serial_number.isin(np.concatenate(train_s_nums))][mask]
    val_data = train_df[train_df.serial_number.isin(val_s_nums)][mask]

    train_data.sample(frac=1)
    val_data.sample(frac=1)
    
    # get X_tr,y_tr,X_val, y_val
    X_tr, y_tr = train_data[mask].drop('RUL_30',axis=1), train_data['RUL_30'] 
    X_vl, y_vl = val_data[mask].drop('RUL_30',axis=1), val_data['RUL_30'] 
    
    # scale the data
    #scaler = StandardScaler()
    #X_tr_scaled = scaler.fit_transform(X_tr)
    #X_vl_scaled = scaler.transform(X_vl)

    # over sample 
    ros = RandomOverSampler(random_state=1)
    X_resampled, y_resampled = ros.fit_sample(X_tr, y_tr)

    # under sample
    #ind = y_tr[y_tr==0].sample(len(y_tr[y_tr==1])).index
    #y_resampled = pd.concat([y_tr[ind],y_tr[y_tr==1]])
    #X_resampled = X_tr.loc[y_resampled.index]
    #y_resampled = y_tr
    #X_resampled = X_tr_scaled
    #X_vl = X_vl_scaled
    # model
    model = RandomForestRegressor(n_estimators=50)
    #model = SGDRegressor()
    model.fit(X_resampled,y_resampled)
    #model = sm.GLM(y_resampled, sm.add_constant(X_resampled), family=sm.families.Poisson())
    #poisson_training_results = model.fit()

    #train_predicted = (model.predict_proba(X_resampled)[:, 1] >= threshold)
    train_predicted = model.predict(X_resampled)
    #val_predicted = (model.predict_proba(X_vl)[:, 1] >= threshold)
    val_predicted = model.predict(X_vl)
    
    
    r2_score_list_tr.append(r2_score(y_resampled, train_predicted))
    
    r2_score_list_vl.append(r2_score(y_vl, val_predicted))

r2_tr_mean  = sum(r2_score_list_tr)/len(r2_score_list_tr)
r2_vl_mean = sum(r2_score_list_vl)/len(r2_score_list_vl)
print(r2_tr_mean)
print(r2_vl_mean)


X_ts, y_ts = test_df[mask].drop('RUL_30',axis=1), test_df['RUL_30'] 

test_pred = model.predict(X_ts)

r2_score(y_ts,test_pred)

We get negative r2 values, whcih mean the models are doing really poorly. One of the reasons could be that the features do not change much overtime, do not degrade gradually, so it's hard to see the difference between 100 remaining useful life and 30.
